init wayland support
The display server based on X or Wayland, is now a choice of different profile in Tizen 3.0.
Basically and consistently, two macros were used "with wayland" and "with x".
Below summarize the combination of the macros:

    | wayland |  x  | meaning
    |---------------------------
    |   0     |  1  | pure X11 platform(no wayland)
    |   1     |  0  | pure wayland platform (no X11)
    |   1     |  1  | wayland but X compatibility
    |   0     |  0  | no X and no wayland

This method unifies the meaning and usage. Deploy this method to syspopup package.

fix the bug:https://bugs.tizen.org/jira/browse/PTREL-531
Signed-off-by: Li,Limin <liminx.li@intel.com>
Change-Id: Id3f61e652e22115d47c7ef8d50dd5cb70b64c112